Fix: Use int instead of float for smoothing window

+++ src/core/libmaven/mavenparameters.h
@@ -126,7 +126,7 @@ public:
 	/**
 	 * smoothing window
 	 */
-	float eic_smoothingWindow;
+	int eic_smoothingWindow;
 	int eic_smoothingAlgorithm;
 	int baseline_smoothingWindow;
 	int baseline_dropTopX;
